# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blood Sanjal API",
    "version": "1.0.0",
    "description": "API for Blood Sanjal platform - A comprehensive blood donation and request management system."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/requests": {
      "post": {
        "tags": [
          "Blood Requests"
        ],
        "summary": "Create a blood request (Normal/Urgent/Emergency)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "patientName": {
                    "type": "string"
                  },
                  "bloodGroup": {
                    "type": "string"
                  },
                  "unitsRequired": {
                    "type": "integer"
                  },
                  "urgency": {
                    "type": "string",
                    "enum": [
                      "NORMAL",
                      "URGENT",
                      "EMERGENCY"
                    ]
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "hospital": {
                        "type": "string"
                      },
                      "province": {
                        "type": "string"
                      },
                      "district": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "example": {
                "patientName": "Jane Doe",
                "bloodGroup": "O+",
                "unitsRequired": 2,
                "urgency": "EMERGENCY",
                "location": {
                  "hospital": "General Hospital",
                  "city": "Kathmandu"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blood request created"
          }
        }
      }
    },
...
    "/donations": {
      "post": {
        "tags": [
          "Donations"
        ],
        "summary": "Record a successful donation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "location": {
                    "type": "string"
                  },
                  "units": {
                    "type": "integer"
                  },
                  "evidenceFileId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Donation recorded successfully"
          }
        }
      }
    },
...
  }
}